+++ Tests/MapboxMapsTests/Foundation/Mocks/MockStyleManager.swift
@@ -701,7 +701,7 @@ struct NonEncodableLayer: Layer {
     var sourceLayer: String?
     var minZoom: Double?
     var maxZoom: Double?
-    var slot: String?
+    var slot: Slot?
 
     init() {}