+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+* Introduce [`Slot`](https://docs.mapbox.com/ios/maps/api/11.0.0-rc.2 -docc/documentation/mapboxmaps/slot) for assining a layer to a slot.
+
 ## 11.0.0-rc.2 - 17 November, 2023
 
 ### Breaking changes ⚠️